refactor: move cog-agnostic helpers into utils
