+++ tests/test_parse_douyin.py
@@ -6,6 +6,7 @@ from pathlib import Path
 
 import pytest
 
+from discordbot.utils import douyin as douyin_fetch
 from discordbot.utils.douyin import (
     DouyinPost,
     DouyinError,
@@ -15,7 +16,6 @@ from discordbot.utils.douyin import (
     DouyinTooLargeError,
     DouyinUnavailableError,
 )
-from discordbot.cogs._parse_douyin import fetch as douyin_fetch
 from discordbot.cogs._gen_reply.link_sources import douyin as douyin_builder
 from discordbot.cogs._gen_reply.link_sources.douyin import (
     DOUYIN_BLOCKED_NOTICE,
